+++ c/expr/unaryop.cc
@@ -5,22 +5,12 @@
 //
 // © H2O.ai 2018
 //------------------------------------------------------------------------------
+#include "expr/base_expr.h"
 #include "expr/py_expr.h"
 #include "types.h"
 
 namespace expr
 {
-
-enum OpCode {
-  IsNa   = 1,
-  Minus  = 2,
-  Plus   = 3,
-  Invert = 4,
-  Abs    = 5,
-  Exp    = 6,
-  LogE   = 7,
-  Log10  = 8,
-};
 
 
 //------------------------------------------------------------------------------
@@ -38,14 +28,14 @@ static void map_n(int64_t row0, int64_t row1, void** params) {
   }
 }
 
-template<typename IT, typename OT, OT (*OP)(IT)>
+template<typename IT, typename OT, OT (*OP)(IT, IT)>
 static void strmap_n(int64_t row0, int64_t row1, void** params) {
   StringColumn<IT>* col0 = static_cast<StringColumn<IT>*>(params[0]);
   Column* col1 = static_cast<Column*>(params[1]);
   const IT* arg_data = col0->offsets();
   OT* res_data = static_cast<OT*>(col1->data_w());
   for (int64_t i = row0; i < row1; ++i) {
-    res_data[i] = OP(arg_data[i]);
+    res_data[i] = OP(arg_data[i - 1] & ~GETNA<IT>(), arg_data[i]);
   }
 }
 
@@ -117,38 +107,57 @@ inline static int8_t bool_inverse(int8_t x) {
 
 
 //------------------------------------------------------------------------------
+// String operators
+//------------------------------------------------------------------------------
+
+template <typename T>
+inline static int8_t op_isna_str(T, T end) {
+  return ISNA<T>(end);
+}
+
+template <typename IT, typename OT>
+inline static OT op_len_str(IT start, IT end) {
+  return ISNA<IT>(end)? GETNA<OT>() : static_cast<OT>(end - start);
+}
+
+
+
+//------------------------------------------------------------------------------
 // Method resolution
 //------------------------------------------------------------------------------
 
 template<typename IT>
-static mapperfn resolve1(int opcode) {
+static mapperfn resolve1(dt::unop opcode) {
   switch (opcode) {
-    case IsNa:    return map_n<IT, int8_t, op_isna<IT>>;
-    case Minus:   return map_n<IT, IT, op_minus<IT>>;
-    case Abs:     return map_n<IT, IT, op_abs<IT>>;
-    case Exp:     return map_n<IT, double, op_exp<IT>>;
-    case LogE:    return map_n<IT, double, op_loge<IT>>;
-    case Log10:   return map_n<IT, double, op_log10<IT>>;
-    case Invert:
+    case dt::unop::ISNA:    return map_n<IT, int8_t, op_isna<IT>>;
+    case dt::unop::MINUS:   return map_n<IT, IT, op_minus<IT>>;
+    case dt::unop::ABS:     return map_n<IT, IT, op_abs<IT>>;
+    case dt::unop::EXP:     return map_n<IT, double, op_exp<IT>>;
+    case dt::unop::LOGE:    return map_n<IT, double, op_loge<IT>>;
+    case dt::unop::LOG10:   return map_n<IT, double, op_log10<IT>>;
+    case dt::unop::INVERT:
       if (std::is_floating_point<IT>::value) return nullptr;
       return map_n<IT, IT, Inverse<IT>::impl>;
+    default:                return nullptr;
   }
-  return nullptr;
 }
+
 
 template<typename T>
-static mapperfn resolve_str(int opcode) {
-  if (opcode == OpCode::IsNa) {
-    return strmap_n<T, int8_t, op_isna<T>>;
+static mapperfn resolve_str(dt::unop opcode) {
+  using OT = typename std::make_signed<T>::type;
+  switch (opcode) {
+    case dt::unop::ISNA: return strmap_n<T, int8_t, op_isna_str<T>>;
+    case dt::unop::LEN:  return strmap_n<T, OT, op_len_str<T, OT>>;
+    default:             return nullptr;
   }
-  return nullptr;
 }
 
 
-static mapperfn resolve0(SType stype, int opcode) {
+static mapperfn resolve0(SType stype, dt::unop opcode) {
   switch (stype) {
     case SType::BOOL:
-      if (opcode == OpCode::Invert) return map_n<int8_t, int8_t, bool_inverse>;
+      if (opcode == dt::unop::INVERT) return map_n<int8_t, int8_t, bool_inverse>;
       return resolve1<int8_t>(opcode);
     case SType::INT8:    return resolve1<int8_t>(opcode);
     case SType::INT16:   return resolve1<int16_t>(opcode);
@@ -164,20 +173,22 @@ static mapperfn resolve0(SType stype, int opcode) {
 }
 
 
-Column* unaryop(int opcode, Column* arg)
+Column* unaryop(dt::unop opcode, Column* arg)
 {
-  if (opcode == OpCode::Plus) return arg->shallowcopy();
+  if (opcode == dt::unop::PLUS) return arg->shallowcopy();
   arg->reify();
 
   SType arg_type = arg->stype();
   SType res_type = arg_type;
-  if (opcode == OpCode::IsNa) {
+  if (opcode == dt::unop::ISNA) {
     res_type = SType::BOOL;
-  } else if (arg_type == SType::BOOL && opcode == OpCode::Minus) {
+  } else if (arg_type == SType::BOOL && opcode == dt::unop::MINUS) {
     res_type = SType::INT8;
-  } else if (opcode == OpCode::Exp || opcode == OpCode::LogE ||
-             opcode == OpCode::Log10) {
+  } else if (opcode == dt::unop::EXP || opcode == dt::unop::LOGE ||
+             opcode == dt::unop::LOG10) {
     res_type = SType::FLOAT64;
+  } else if (opcode == dt::unop::LEN) {
+    res_type = arg_type == SType::STR32? SType::INT32 : SType::INT64;
   }
   void* params[2];
   params[0] = arg;
@@ -186,7 +197,7 @@ Column* unaryop(int opcode, Column* arg)
   mapperfn fn = resolve0(arg_type, opcode);
   if (!fn) {
     throw RuntimeError()
-      << "Unable to apply unary op " << opcode << " to column(stype="
+      << "Unable to apply unary op " << int(opcode) << " to column(stype="
       << arg_type << ")";
   }