+++ datatable/expr/__init__.py
@@ -28,11 +28,10 @@ from .cast_expr import CastExpr
 from .column_expr import ColSelectorExpr, NewColumnExpr
 from .dtproxy import f, g
 from .exp_expr import exp, log, log10
-from .isna_expr import isna
 from .literal_expr import LiteralExpr
 from .reduce_expr import ReduceExpr, sum, count, first, mean, min, max, sd
 from .relop_expr import RelationalOpExpr
-from .unary_expr import UnaryOpExpr
+from .unary_expr import UnaryOpExpr, isna
 
 __all__ = (
     "abs",