+++ datatable/expr/base_expr.py
@@ -5,8 +5,6 @@
 #   file, You can obtain one at http://mozilla.org/MPL/2.0/.
 #-------------------------------------------------------------------------------
 import datatable
-from .consts import ctypes_map, nas_map
-
 
 
 
@@ -186,6 +184,8 @@ class BaseExpr:
         """Unary plus (no-op)."""
         return datatable.expr.UnaryOpExpr("+", self)
 
+    def len(self):
+        return datatable.expr.UnaryOpExpr("len", self)
 
 
     #----- Code generation -----------------------------------------------------
@@ -210,6 +210,3 @@ class BaseExpr:
     def __repr__(self):
         return "%s(%s)" % (self.__class__.__name__, self)
 
-
-    def safe_name(self):
-        return str(self)