+++ datatable/expr/cast_expr.py
@@ -4,12 +4,13 @@
 #   License, v. 2.0. If a copy of the MPL was not distributed with this
 #   file, You can obtain one at http://mozilla.org/MPL/2.0/.
 #-------------------------------------------------------------------------------
-
 from .base_expr import BaseExpr
-from .consts import baseexpr_opcodes
 from datatable.lib import core
 
 __all__ = ("CastExpr", )
+
+# See "c/expr/base_expr.h"
+BASEEXPR_OPCODE_CAST = 5
 
 
 
@@ -25,6 +26,6 @@ class CastExpr(BaseExpr):
         return "%s(%s)" % (self._stype.name, str(self._arg))
 
     def _core(self):
-        return core.base_expr(baseexpr_opcodes["cast"],
+        return core.base_expr(BASEEXPR_OPCODE_CAST,
                               self._arg._core(),
                               self._stype.value)